requirements.txt
dash==2.18.2
plotly==5.24.1
databricks-sql-connector==3.6.0
databricks-sdk==0.38.0
pyyaml==6.0.2
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
shap==0.46.0
mlflow==2.19.0
prophet==1.1.6
xgboost==2.1.3
